add flutter app and organized all java programs
